Add files via upload
Sultan ul Ashiqeen founded Tehreek Dawat e Faqr on October 23, 2009. Tehreek Dawat-e-Faqr is the only representative of the teachings of Faqr. It is a registered organization, completely non-political and non-governmental. Furthermore, it is non-sectarian.
Our aim is the revival of a strong spiritual relationship between Allah and Muslims.  Deteriorated spiritual relations have resulted in the decline of Muslims despite the fact it is the best umma. With the teachings of Faqr, we aim at redirecting Muslims from physical devotions of pretence to sanctification of soul. Thus making physical devotions acceptable in the court of Allah.